feat: Room Database에 예매 내역을 저장하는 로직 구현
