+++ app/src/main/java/woowacourse/movie/data/entity/TicketEntity.kt
@@ -6,7 +6,6 @@ import androidx.room.PrimaryKey
 import woowacourse.movie.domain.model.HeadCount
 import woowacourse.movie.domain.model.Movie
 import woowacourse.movie.domain.model.Theater
-import woowacourse.movie.domain.model.Ticket
 import woowacourse.movie.domain.model.seat.Seats
 import java.time.LocalDateTime
 
@@ -19,14 +18,5 @@ data class TicketEntity(
     @ColumnInfo(name = "seats") val seats: Seats,
 ) {
     @PrimaryKey(autoGenerate = true)
-    var uid: Long = 0
+    var id: Long = 0
 }
-
-fun TicketEntity.toDomainModel(): Ticket =
-    Ticket(
-        movie = this.movie,
-        theater = this.theater,
-        showtime = this.showtime,
-        headCount = this.headCount,
-        seats = this.seats,
-    )
