+++ app/src/main/java/woowacourse/movie/data/repository/TicketRepository.kt
@@ -2,21 +2,30 @@ package woowacourse.movie.data.repository
 
 import woowacourse.movie.data.database.MovieDatabase
 import woowacourse.movie.data.entity.toDomainModel
+import woowacourse.movie.data.entity.toEntity
 import woowacourse.movie.domain.model.Ticket
 import kotlin.concurrent.thread
 
 interface TicketRepository {
-    fun getTickets(): List<Ticket>
+    fun getAll(): List<Ticket>
+
+    fun save(ticket: Ticket)
 }
 
 class DefaultTicketRepository(
     private val database: MovieDatabase,
 ) : TicketRepository {
-    override fun getTickets(): List<Ticket> {
+    override fun getAll(): List<Ticket> {
         var result: List<Ticket> = emptyList()
         thread {
             result = database.ticketDao.getAll().map { it.toDomainModel() }
         }.join()
         return result
+    }
+
+    override fun save(ticket: Ticket) {
+        thread {
+            database.ticketDao.insert(ticket.toEntity())
+        }.join()
     }
 }
